Show parse tree for regex using treant-js
Unfortunately the npm version of treant-js was buggy,
but the version on their website worked, included in this commit.

I couldn't get the Regex.rei to work properly, so deleted it
until I figure out how to do it properly.

I would also like to extract the regexp2parseTree function from Regex.re
so that I can continue to auto-convert new code from the OCaml project,
but when I put regexp2parseTree in a separate file it didn't work,
probably due to not finding the types in Regex.re.

At least now it works, but looking forward to clean this mess up later.