Fix bug causing negated bracket expressions to fail if overlapping
I discovered this bug when I got error "%01-t overlaps a-q"
when testing the regex "[a-q][^u-z]".

The problem was in the construction of the RangeSet for negated bracket expressions,
which are created by converting to/from CharSet, and taking the diff between
a CharSet containing all characters, and the RangeSet for the negated bracket expression.

The default was allow_overlap=false in the of_char_set function,
but at this stage during parsing, we need to allow overlaps.

Change default allow_overlap=true, which is correct, since this should be set to false
in the final step after ranges have been factorized.